statistics/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
astroid==3.2.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.9.0
cachetools==5.3.2
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.8
contourpy==1.3.2
coverage==7.3.2
cycler==0.12.1
Deprecated==1.2.15
dill==0.3.8
discord==2.3.2
discord.py==2.4.0
distlib==0.3.7
exif==1.6.0
filelock==3.13.1
flake8==6.1.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Limiter==3.9.2
fonttools==4.46.0
frozendict==2.4.6
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.10
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.37.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.2
google-cloud-recaptcha-enterprise==1.26.0
google-genai==1.11.0
google-generativeai==0.8.3
googleapis-common-protos==1.61.0
grpcio==1.68.1
grpcio-status==1.62.3
gspread==6.2.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
identify==2.5.32
idna==3.6
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
japanize-matplotlib==1.1.3
Jinja2==3.1.5
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.8
limits==3.14.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
mdurl==0.1.2
multidict==6.1.0
multitasking==0.0.11
mypy==1.10.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
narwhals==2.5.0
nodeenv==1.8.0
numpy==2.2.4
oauthlib==3.3.1
ordered-set==4.1.0
packaging==23.2
pandas==2.2.3
# Editable install with no version control (pdf-annotate==0.12.0)
-e /Users/keita/projects/pdfrw/client-projects/pdf-annotate
# Editable install with no version control (pdfrw==0.4)
-e /Users/keita/projects/pdfrw
peewee==3.17.9
Pillow==10.1.0
platformdirs==4.1.0
pluggy==1.5.0
plum-py==0.8.7
pre-commit==3.5.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==4.25.1
pyarrow==21.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
pyflakes==3.1.0
Pygments==2.18.0
pylint==3.2.3
pyparsing==3.1.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-geohash==0.8.5
pytz==2024.1
PyYAML==6.0.1
referencing==0.36.2
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.27.1
rsa==4.9
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.5
streamlit==1.49.1
tenacity==9.1.2
toml==0.10.2
tomlkit==0.12.5
tornado==6.5.2
tqdm==4.67.1
types-requests==2.32.0.20240622
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.25.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.18.3
yfinance==0.2.55
